# resources/api-doc.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/instructors/{id}:
    get:
      tags:
        - instructor-controller
      operationId: getInstructorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Instructor'
    put:
      tags:
        - instructor-controller
      operationId: updateInstructor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Instructor'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Instructor'
  /api/instructors:
    post:
      tags:
        - instructor-controller
      operationId: createInstructor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Instructor'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Instructor'
  /api/attendance/check-out:
    post:
      tags:
        - attendance-controller
      operationId: checkOut
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorIdRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CheckOutResponse'
  /api/attendance/check-in:
    post:
      tags:
        - attendance-controller
      operationId: checkIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorIdRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CheckInResponse'
  /api/reports/monthly:
    get:
      tags:
        - reports-controller
      operationId: getMonthlyAttendance
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            enum:
              - JANUARY
              - FEBRUARY
              - MARCH
              - APRIL
              - MAY
              - JUNE
              - JULY
              - AUGUST
              - SEPTEMBER
              - OCTOBER
              - NOVEMBER
              - DECEMBER
        - name: year
          in: query
          required: true
          schema:
            type: string
        - name: pageNo
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageWrapper'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        description:
          type: string
    Instructor:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    InstructorIdRequest:
      required:
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
    CheckOutResponse:
      type: object
      properties:
        attendanceId:
          type: integer
          format: int64
        checkOutTime:
          type: string
          format: date-time
        instructorId:
          type: integer
          format: int64
    CheckInResponse:
      type: object
      properties:
        attendanceId:
          type: integer
          format: int64
        checkInTime:
          type: string
          format: date-time
        instructorId:
          type: integer
          format: int64
    AttendanceResponse:
      type: object
      properties:
        instructorId:
          type: integer
          format: int64
        timeInMinutes:
          type: integer
          format: int64
    PageWrapper:
      type: object
      properties:
        totalInstructors:
          type: integer
          format: int64
        instructors:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceResponse'
        currentPage:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32